jest and enzyme working
