+++ jest.config.js
@@ -1,3 +1,3 @@
 module.exports = {
-  setupFiles: ['<rootDir>/spec/shim.js', '<rootDir>/spec/setUpTests.js']
+  setupFiles: ['<rootDir>/spec/setUpTests.js']
 }
