+++ package.json
@@ -34,7 +34,10 @@
     "eslint-plugin-import": "^2.14.0",
     "eslint-plugin-jsx-a11y": "^6.1.2",
     "eslint-plugin-react": "^7.11.1",
+    "install": "^0.12.2",
     "jest": "^23.6.0",
+    "npm": "^6.4.1",
+    "react-test-render": "^1.1.1",
     "regenerator-runtime": "^0.12.1",
     "webpack": "^4.23.0",
     "webpack-cli": "^3.1.2"
