+++ spec/setUpTests.js
@@ -1,5 +1,3 @@
-// import React from 'react';
-// import { shallow, mount, render } from 'enzyme';
 
   import Enzyme from 'enzyme';
   import Adapter from 'enzyme-adapter-react-16';
@@ -7,3 +5,5 @@
 
 
   Enzyme.configure({ adapter: new Adapter() });
+
+  export default Enzyme;
